pyproject.toml
[project]
name = "project2"
version = "0.1.0"
description = "Traffic Light Autoencoder - CS 4774 Project 2"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["infrastructure", "models"]
